ACP: split control-plane manager into core/types/utils modules
